Activiti basics part 2 - making a dynamic workflow.
- Publish the workflow via Activiti console
- Pick any file through the Share UI logged in as abeecher (may need to activate account / reset password)
- Start workflow with the file, assign it to admin for approval
- See the workflow live via Process instances on Activiti console
- Approve and reject to your heart's content - note that abeecher is the originator and admin the approver (twice)

This part adds proper responses to approve/reject and cares about who was assigned to the workflow. Initial user submits, revises or discards if needed, and publishes at the end. Assigned user reviews, getting two chances to reject (second could be set to another user)

Listeners are needed on the review steps to fetch the result from the task. All three are "Alfresco task scripts" and the event picked MUST be set to "complete" (so on completion - defaults to "create" at which point the task was created but the result hasn't been submitted yet)

Along with the listeners come conditions on the approve/reject legs from the gateways so the proper leg is picked to match the answer.