Next stage of fanciness: Automatically apply (replacing any existing) category as a document travels through a workflow.
Includes a few new categories including end states, but need to review what needs to listen differently there. Category parents don't list the contents of their children :-(

Introduces the use of fields on listeners for User Tasks in the workflow and injecting the field value to the listener.

Could be improved further such as with cached/static categories, avoiding (or supporting) multiple category hits, etc.