Auto-publish workflows on startup.
Note that this publishes *all* the listed workflows on *every* startup, resulting in a new version each time. New actions use the latest version and only one appears in "Deployed process definitions" while "Deployments" shows the duplicate older versions.

Existing process instances should in theory keep going on the path of their old version of the workflow. Important to keep this in mind for design. Easiest is probably either using this during development to avoid having to constantly republish manually and for cases in which all workflows expire naturally and none are kept around on older versions (risking confusion and bugs) during releases.

The "redeploy" property in the bootstrap file indicates whether a workflow should be deployed even if it already exists. However, perhaps counter-intuitively, it does not actually replace the old copy. If the property is false the workflow is simply not deployed if a copy already exists.